#10077: Overload binary gt in unary gt
