--- docs/source/ttnn/ttnn/ttnn/binary_gt_bw.rst
@@ -1,6 +1,0 @@
-.. _ttnn.binary_gt_bw:
-
-ttnn.binary_gt_bw
-#################
-
-.. autofunction:: ttnn.binary_gt_bw
