--- tests/ttnn/unit_tests/operations/backward/test_backward_binary_gt.py
@@ -1,28 +1,0 @@
-# SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
-
-# SPDX-License-Identifier: Apache-2.0
-
-import torch
-import pytest
-import ttnn
-from tests.ttnn.unit_tests.operations.backward.utility_funcs import data_gen_with_range, compare_pcc
-
-
-@pytest.mark.parametrize(
-    "input_shapes",
-    (
-        (torch.Size([1, 1, 32, 32])),
-        (torch.Size([1, 1, 320, 384])),
-        (torch.Size([1, 3, 320, 384])),
-    ),
-)
-def test_bw_binary_gt(input_shapes, device):
-    in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
-    other_data, other_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
-    grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 100, device)
-
-    tt_output_tensor_on_device = ttnn.binary_gt_bw(grad_tensor, input_tensor, other_tensor)
-    pt_y = torch.zeros_like(grad_data)
-    golden_tensor = [pt_y, pt_y]
-    comp_pass = compare_pcc(tt_output_tensor_on_device, golden_tensor)
-    assert comp_pass
