+++ tests/ttnn/unit_tests/operations/backward/test_backward_gt.py
@@ -16,6 +16,26 @@ from tests.ttnn.unit_tests.operations.backward.utility_funcs import data_gen_wit
         (torch.Size([1, 3, 320, 384])),
     ),
 )
+def test_bw_binary_gt(input_shapes, device):
+    in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
+    other_data, other_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
+    grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 100, device)
+
+    tt_output_tensor_on_device = ttnn.gt_bw(grad_tensor, input_tensor, other_tensor)
+    pt_y = torch.zeros_like(grad_data)
+    golden_tensor = [pt_y, pt_y]
+    comp_pass = compare_pcc(tt_output_tensor_on_device, golden_tensor)
+    assert comp_pass
+
+
+@pytest.mark.parametrize(
+    "input_shapes",
+    (
+        (torch.Size([1, 1, 32, 32])),
+        (torch.Size([1, 1, 320, 384])),
+        (torch.Size([1, 3, 320, 384])),
+    ),
+)
 @pytest.mark.parametrize("other", [1.0])
 def test_bw_unary_gt(input_shapes, other, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
